refactored as test framework cannot yet support subpackage testing
