+++ internal/services/appservice/service_plan_resource_test.go
@@ -1,4 +1,4 @@
-package serviceplan_test
+package appservice_test
 
 import (
 	"context"
