+++ internal/services/appservice/source_control_resource.go
@@ -1,4 +1,4 @@
-package sourcecontrol
+package appservice
 
 import (
 	"context"
