+++ internal/services/appservice/source_control_resource_test.go
@@ -1,4 +1,4 @@
-package sourcecontrol_test
+package appservice_test
 
 import (
 	"context"
