+++ internal/services/appservice/source_control_token_data_source.go
@@ -1,4 +1,4 @@
-package sourcecontrol
+package appservice
 
 import (
 	"context"
