+++ internal/services/appservice/web_app_fx_strings.go
@@ -1,4 +1,4 @@
-package webapp
+package appservice
 
 import (
 	"strings"
